fixed bug in Config View Helper
the Config View Helper provides the possibility to access Configuration in view scripts
even it's worth to discuss if config access should be available within view scripts it makes sense in some situations.
Espespecially to avoid duplication of scripts for different views if these scripts differentiate only very little.

implemented by snowflake the helper couldn't be used as it is.