Mark CitationTest skipped if short_open_tags is disabled
Mark test skipped if short_open_tag is not enabled. The partial uses
short open tags. This directive is PHP_INI_PERDIR, i.e. can only be
changed via php.ini or a per-directory equivalent. The testCitations()
will fail if the test is run on a system with short_open_tag disabled
in the system-wide php ini-file.