Add schema.org structured data for bibs and holdings
Add basic RDFa Lite markup for bibliographic attributes to the default
record template.

Following the emergined recommendations of the W3C Schema Bib Extend
community group around individual item holdings, mark up each available
item as an instance of http://schema.org/Offer to be as compatible as
possible with how online vendors mark up their offerings. Due to the
commercial orientation of Offer, the mappings may look a little odd, but
they are in the spirit of the defined attributes.

* library = Offer/seller
* callnumber = Offer/sku
* barcode = Offer/serialNumber
* status = Offer/itemAvailability

Signed-off-by: Dan Scott <dan@coffeecode.net>